requirements.txt
asgiref==3.8.1
attrs==23.2.0
backports.zoneinfo==0.2.1
black==24.4.2
click==8.1.7
Django==4.2.13
django-faker==0.2
djangorestframework==3.15.1
fake-factory==9999.9.9
Faker==25.8.0
flake8==7.0.0
importlib-resources==6.4.0
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pkgutil-resolve-name==1.3.10
platformdirs==4.2.2
pycodestyle==2.11.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
typing-extensions==4.12.2
uritemplate==4.1.1
zipp==3.19.2
